+++ packages/styled-components/src/native/test/native.test.js
@@ -1,5 +1,5 @@
 // @flow
-/* eslint-disable react/prop-types */
+import 'react-native';
 import { Text, View } from 'react-native';
 import React from 'react';
 import TestRenderer from 'react-test-renderer';
@@ -77,43 +77,6 @@ describe('native', () => {
     expect(view.props.style).toEqual([{ color: 'red', textAlign: 'left' }]);
   });
 
-  it('folds defaultProps', () => {
-    const Inner = styled.View``;
-
-    Inner.defaultProps = {
-      theme: {
-        fontSize: 12,
-      },
-      style: {
-        background: 'blue',
-        textAlign: 'center',
-      },
-    };
-
-    const Outer = styled(Inner)``;
-
-    Outer.defaultProps = {
-      theme: {
-        fontSize: 16,
-      },
-      style: {
-        background: 'silver',
-      },
-    };
-
-    expect(Outer.defaultProps).toMatchInlineSnapshot(`
-Object {
-  "style": Object {
-    "background": "silver",
-    "textAlign": "center",
-  },
-  "theme": Object {
-    "fontSize": 16,
-  },
-}
-`);
-  });
-
   it('should combine inline styles and the style prop', () => {
     const Comp = styled.View`
       padding-top: 10;
@@ -154,18 +117,6 @@ Object {
     wrapper.update(<Comp opacity={0.9} />);
 
     expect(wrapper.root.findByType('View').props.style).toEqual([{ paddingTop: 5, opacity: 0.9 }]);
-  });
-
-  it('should pass "forwardedAs" to the underlying component as "as" if used', () => {
-    const Comp = ({ as: Component = View, ...props }) => <Component {...props} />;
-
-    const Comp2 = styled(Comp)`
-      background: red;
-    `;
-
-    const wrapper = TestRenderer.create(<Comp2 forwardedAs={Text} />);
-
-    expect(wrapper.root.findByType('Text')).not.toBeUndefined();
   });
 
   describe('attrs', () => {
@@ -299,33 +250,29 @@ Object {
     });
 
     it('should override children', () => {
-      const child = <Text>Amazing</Text>;
-
       const Comp = styled.Text.attrs({
-        children: child,
+        children: <Text>Amazing</Text>,
       })``;
 
       const wrapper = TestRenderer.create(<Comp>Something else</Comp>);
       const text = wrapper.root.findByType('Text');
 
       expect(text.props).toMatchObject({
-        children: child,
+        children: 'Something else',
         style: [{}],
       });
     });
 
     it('accepts a function', () => {
-      const child = <Text>Amazing</Text>;
-
-      const Comp = styled.Text.attrs(() => ({
-        children: child,
+      const Comp = styled.Text.attrs(props => ({
+        children: <Text>Amazing</Text>,
       }))``;
 
       const wrapper = TestRenderer.create(<Comp>Something else</Comp>);
       const text = wrapper.root.findByType('Text');
 
       expect(text.props).toMatchObject({
-        children: child,
+        children: 'Something else',
         style: [{}],
       });
     });
@@ -351,25 +298,33 @@ Object {
 
     it('theme prop works', () => {
       const Comp = styled.Text`
-        color: ${({ theme }) => theme.myColor};
+        color: ${({theme}) => theme.myColor};
       `;
 
-      const wrapper = TestRenderer.create(<Comp theme={{ myColor: 'red' }}>Something else</Comp>);
+      const wrapper = TestRenderer.create(
+        <Comp theme={{myColor: 'red'}}>Something else</Comp>
+      );
       const text = wrapper.root.findByType('Text');
 
-      expect(text.props.style).toMatchObject([{ color: 'red' }]);
+      expect(text.props.style).toMatchObject(
+        [{"color": "red"}],
+      );
     });
 
     it('theme in defaultProps works', () => {
       const Comp = styled.Text`
-        color: ${({ theme }) => theme.myColor};
+        color: ${({theme}) => theme.myColor};
       `;
-      Comp.defaultProps = { theme: { myColor: 'red' } };
+      Comp.defaultProps = {theme: {myColor: 'red'}}
 
-      const wrapper = TestRenderer.create(<Comp>Something else</Comp>);
+      const wrapper = TestRenderer.create(
+        <Comp>Something else</Comp>
+      );
       const text = wrapper.root.findByType('Text');
 
-      expect(text.props.style).toMatchObject([{ color: 'red' }]);
+      expect(text.props.style).toMatchObject(
+        [{"color": "red"}],
+      );
     });
   });
 
@@ -466,7 +421,9 @@ For example, { component: () => InnerComponent } instead of { component: InnerCo
       expect(console.warn.mock.calls[0][0]).toMatchInlineSnapshot(
         `"Functions as object-form attrs({}) keys are now deprecated and will be removed in a future version of styled-components. Switch to the new attrs(props => ({})) syntax instead for easier and more powerful composition. The attrs key in question is \\"data-text-color\\" on component \\"Styled(View)\\"."`
       );
-      expect(console.warn.mock.calls[0][1]).toEqual(expect.stringMatching(/^\s+Error\s+at/));
+      expect(console.warn.mock.calls[0][1]).toEqual(
+        expect.stringMatching(/^\s+Error\s+at/)
+      );
     });
   });
 });